+++ owncloudApp/src/main/java/com/owncloud/android/presentation/ui/settings/fragments/SettingsMoreFragment.kt
@@ -20,7 +20,6 @@
 
 package com.owncloud.android.presentation.ui.settings.fragments
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
@@ -29,7 +28,6 @@ import com.owncloud.android.BuildConfig
 import com.owncloud.android.R
 import com.owncloud.android.extensions.goToUrl
 import com.owncloud.android.extensions.sendEmail
-import com.owncloud.android.presentation.ui.settings.PrivacyPolicyActivity
 import com.owncloud.android.presentation.viewmodels.settings.SettingsMoreViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -43,7 +41,6 @@ class SettingsMoreFragment : PreferenceFragmentCompat() {
     private var prefSync: Preference? = null
     private var prefRecommend: Preference? = null
     private var prefFeedback: Preference? = null
-    private var prefPrivacyPolicy: Preference? = null
     private var prefImprint: Preference? = null
 
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
@@ -54,7 +51,6 @@ class SettingsMoreFragment : PreferenceFragmentCompat() {
         prefSync = findPreference(PREFERENCE_SYNC_CALENDAR_CONTACTS)
         prefRecommend = findPreference(PREFERENCE_RECOMMEND)
         prefFeedback = findPreference(PREFERENCE_FEEDBACK)
-        prefPrivacyPolicy = findPreference(PREFERENCE_PRIVACY_POLICY)
         prefImprint = findPreference(PREFERENCE_IMPRINT)
 
         // Help
@@ -109,17 +105,6 @@ class SettingsMoreFragment : PreferenceFragmentCompat() {
             moreScreen?.removePreference(prefFeedback)
         }
 
-        // Privacy policy
-        if (moreViewModel.isPrivacyPolicyEnabled()) {
-            prefPrivacyPolicy?.setOnPreferenceClickListener {
-                val intent = Intent(context, PrivacyPolicyActivity::class.java)
-                startActivity(intent)
-                true
-            }
-        } else {
-            moreScreen?.removePreference(prefPrivacyPolicy)
-        }
-
         // Imprint
         if (moreViewModel.isImprintEnabled()) {
             prefImprint?.setOnPreferenceClickListener {
@@ -138,7 +123,6 @@ class SettingsMoreFragment : PreferenceFragmentCompat() {
         private const val PREFERENCE_SYNC_CALENDAR_CONTACTS = "syncCalendarContacts"
         private const val PREFERENCE_RECOMMEND = "recommend"
         private const val PREFERENCE_FEEDBACK = "feedback"
-        private const val PREFERENCE_PRIVACY_POLICY = "privacyPolicy"
         private const val PREFERENCE_IMPRINT = "imprint"
     }